requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
scikit-learn>=0.24.0
Pillow>=8.0.0
numpy>=1.21.0
tqdm>=4.62.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
gradio>=3.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
